requirements.txt
uvicorn~=0.34.0
fastapi~=0.115.8
pydantic~=2.10.6
h11~=0.14.0
pip~=25.0
sniffio~=1.3.1
typing_extensions~=4.12.2
starlette~=0.45.3
pydantic_core~=2.27.2
anyio~=4.8.0
click~=8.1.8
annotated-types~=0.7.0
pandas~=2.2.3
numpy~=2.2.2
PyYAML~=6.0.2
python-multipart~=0.0.20